vertical merge tests update; docs
