+++ tests/test_data.py
@@ -1760,65 +1760,6 @@ class TestOfflineDataSetup(unittest.TestCase):
         Annotation(document=document, spans=[first_span, second_span], label_set=self.label_set, label=self.label)
         assert len(document.annotations(use_correct=False)) == 2
 
-    def test_merge_vertical_1(self):
-        """Test the vertical merging of Spans into a single Annotation."""
-        project = LocalTextProject()
-
-        category = project.get_category_by_id(1)
-
-        document = project.no_status_documents[1]
-        label = document.annotations(use_correct=False)[0].label
-
-        assert len(document.spans()) == 4
-        assert len(document.annotations(use_correct=False)) == 4
-
-        with pytest.raises(TypeError, match="This value has never been computed."):
-            document.merge_vertical()
-
-        for category_label in category.labels:
-            category_label.has_multiline_annotations(categories=[category])
-
-        document.merge_vertical()
-
-        assert label.has_multiline_annotations(categories=[category]) is False
-        assert document.bboxes_available is True
-
-        train_document = Document(project=project, category=category, text='p1\np2', dataset_status=2)
-        train_span1 = Span(start_offset=0, end_offset=2)
-        train_span2 = Span(start_offset=3, end_offset=5)
-        _ = Annotation(
-            document=train_document,
-            is_correct=True,
-            label=label,
-            label_set=project.no_label_set,
-            spans=[train_span1, train_span2],
-        )
-
-        assert label.has_multiline_annotations() is False  # Value hasn't been updated yet
-        assert label.has_multiline_annotations(categories=[category]) is True
-        assert label.has_multiline_annotations() is True
-
-        document.merge_vertical()
-
-        assert len(document.spans()) == 4
-        assert len(document.annotations(use_correct=False)) == 2
-
-    def test_merge_vertical_2(self):
-        """Test the vertical merging of Spans into a single Annotation."""
-        project = LocalTextProject()
-
-        document = project.get_document_by_id(8)
-
-        assert len(document.annotations(use_correct=False)) == 6
-
-        with pytest.raises(TypeError, match="This value has never been computed."):
-            document.merge_vertical(only_multiline_labels=True)
-
-        document.merge_vertical(only_multiline_labels=False)
-
-        assert len(document.annotations(use_correct=False)) == 4
-        assert len(document.annotations(use_correct=False)[1].spans) == 3
-
     def test_lose_weight(self):
         """Lose weight should remove session and Documents."""
         project = Project(id_=None)
