+++ tests/trainer/test_information_extraction.py
@@ -19,7 +19,7 @@ import pandas as pd
 from sklearn.datasets import make_classification
 from sklearn.ensemble import RandomForestClassifier
 
-from konfuzio_sdk.data import Project, Document, AnnotationSet
+from konfuzio_sdk.data import Project, Document, AnnotationSet, Annotation, Span
 from konfuzio_sdk.trainer.information_extraction import (
     num_count,
     date_count,
@@ -764,6 +764,78 @@ class TestInformationExtraction(unittest.TestCase):
         pipeline = RFExtractionAI(category=category)
         res_dict = pipeline.extract_template_with_clf("doc text", res_dict={})
         assert res_dict == {}
+
+    def test_merge_vertical_1(self):
+        """Test the vertical merging of Spans into a single Annotation."""
+        project = LocalTextProject()
+
+        category = project.get_category_by_id(1)
+        pipeline = RFExtractionAI()
+
+        document = project.no_status_documents[1]
+        label = document.annotations(use_correct=False)[0].label
+
+        assert len(document.spans()) == 4
+        assert len(document.annotations(use_correct=False)) == 4
+
+        with pytest.raises(AttributeError, match="merge_vertical requires a Category"):
+            pipeline.merge_vertical(document, only_multiline_labels=True)
+
+        pipeline.category = category
+
+        with pytest.raises(TypeError, match="This value has never been computed."):
+            pipeline.merge_vertical(document)
+
+        for category_label in category.labels:
+            category_label.has_multiline_annotations(categories=[category])
+
+        pipeline.merge_vertical(document)
+
+        assert label.has_multiline_annotations(categories=[category]) is False
+        assert document.bboxes_available is True
+
+        train_document = Document(project=project, category=category, text='p1\np2', dataset_status=2)
+        train_span1 = Span(start_offset=0, end_offset=2)
+        train_span2 = Span(start_offset=3, end_offset=5)
+        _ = Annotation(
+            document=train_document,
+            is_correct=True,
+            label=label,
+            label_set=project.no_label_set,
+            spans=[train_span1, train_span2],
+        )
+
+        assert label.has_multiline_annotations() is False  # Value hasn't been updated yet
+        assert label.has_multiline_annotations(categories=[category]) is True
+        assert label.has_multiline_annotations() is True
+
+        pipeline.merge_vertical(document)
+
+        assert len(document.spans()) == 4
+        assert len(document.annotations(use_correct=False)) == 2
+
+    def test_merge_vertical_2(self):
+        """Test the vertical merging of Spans into a single Annotation with the only_multiline_labels parameter."""
+        project = LocalTextProject()
+        category = project.get_category_by_id(1)
+        pipeline = RFExtractionAI()
+
+        document = project.get_document_by_id(8)
+
+        assert len(document.annotations(use_correct=False)) == 6
+
+        with pytest.raises(AttributeError, match="merge_vertical requires a Category"):
+            pipeline.merge_vertical(document, only_multiline_labels=True)
+
+        pipeline.category = category
+
+        with pytest.raises(TypeError, match="This value has never been computed."):
+            pipeline.merge_vertical(document, only_multiline_labels=True)
+
+        pipeline.merge_vertical(document, only_multiline_labels=False)
+
+        assert len(document.annotations(use_correct=False)) == 4
+        assert len(document.annotations(use_correct=False)[1].spans) == 3
 
     def test_feature_function(self):
         """Test to generate features."""
